Move DeprecatedAttribute annotation to internal package
